refactor: split core/sdk
