+++ src/core/cache/manager.py
@@ -1,5 +1,5 @@
-from src.sdk.cache import raw_db, mint_db, cursor_db
 from pymongo.errors import BulkWriteError
+from . import raw_db, mint_db, cursor_db
 from ..exception import EmptyCache
 from ..logger import log
