+++ src/core/exception.py
