+++ src/core/media/assets/image.py
